[PBW-4586] refactor closed caption screen to use content screen
