--- scss/components/_closed-captions-screen.scss
@@ -1,200 +1,0 @@
-.oo-closed-captions-screen {
-  height: 100%;
-  width: 100%;
-
-  .oo-closed-captions-panel {
-    padding: $panel-padding;
-    bottom: 0;
-    background-color: $panel-bg;
-    left: 0;
-    right: 0;
-    height: 100%;
-
-    .oo-closed-captions-panel-title {
-      position: relative;
-      line-height: 1;
-      font-weight: bold;
-      font-size: $font-size-md-screen-title;
-      font-family: $font-family-roboto-condensed;
-      span.oo-icon {
-        top: 1px;
-        position: relative;
-        margin-left: 7px;
-        font-size: 0.8em;
-      }
-    }
-
-    .oo-switch-container {
-      cursor: pointer;
-      margin-top: 35px;
-      height: 21px;
-      width: 105px;
-      position: relative;
-
-      .oo-switch-container-selectable {
-        @extend .oo-expand;
-        text-decoration:none;
-        cursor: pointer;
-        //for IE
-        background-color:$white;
-        opacity: 0;
-        filter: alpha(opacity=1);
-      }
-      .oo-switch-captions {
-        position: absolute;
-        display: inline-block;
-        color: $gray;
-        font-size: $font-size-base;
-        font-family: $font-family-roboto;
-      }
-      .oo-switch-captions-on {
-        right: 0;
-      }
-      .oo-switch-captions-off {
-        left: 0;
-      }
-      .oo-switch-captions-active {
-        color: $white;
-      }
-      .oo-switch-element {
-        display: inline-block;
-        width: 33%;
-        height: 100%;
-        position: relative;
-        margin-left: 33%;
-        border-radius: 100%;
-      }
-      .oo-switch-thumb {
-        display: block;
-        width: 60%;
-        height: 100%;
-        position: absolute;
-        top: 0;
-        background: $white;
-        border-radius: 100%;
-        transition: all 0.1s ease-in 0s;
-      }
-      .oo-switch-thumb-on {
-        right: 0;
-      }
-      .oo-switch-thumb-off {
-        left: 0;
-      }
-      .oo-switch-body {
-        display: block;
-        width: 100%;
-        height: 75%;
-        position: absolute;
-        right: 0;
-        top: 12.5%;
-        border-radius: 40px;
-        transition: all 0.5s ease-in-out;
-        background: $blue-light;
-      }
-      .oo-switch-body-off {
-        background: $gray;
-      }
-    }
-
-    .oo-language-container {
-      @include vertical-center;
-      left: 0;
-      right: 0;
-
-      .oo-language-panel {
-        background-color: transparent;
-        min-height: 54%;
-        height: 84%;
-        width: 90%;
-        overflow: auto;
-        white-space: nowrap;
-        margin: auto;
-
-        a.oo-item {
-          width: 140px;
-          font-size: 1.2em;
-          font-family: $font-family-roboto-condensed;
-          text-overflow: ellipsis;
-          white-space: nowrap;
-          overflow: hidden;
-          text-align: center;
-          margin: 14px 28px;
-          padding: 3px;
-          color: $white;
-          transition: all 0.0s ease;
-          cursor: pointer;
-          border-radius: 2px;
-          border: 1px solid $gray;
-          outline: none;
-        }
-
-        a.oo-item-selected {
-          background-color: $blue-light;
-          border-radius: 4px;
-        }
-
-        a.oo-disabled {
-          color: $gray;
-        }
-      }
-
-      .oo-flexcontainer {
-        display: -webkit-flex;
-        display: flex;
-        -webkit-flex-flow: row nowrap;
-        flex-flow: row nowrap;
-        overflow: visible;
-        align-items: center;
-        -webkit-align-items: center;
-        justify-content: center;
-      }
-
-      .oo-discover-btn {
-        position: absolute;
-        top: 48%;
-        background-color: transparent;
-        height: 10%;
-        width: 30px;
-        color: white;
-      }
-
-      .oo-left-button {
-        @extend .oo-discover-btn;
-        left: 0;
-        text-align: right;
-      }
-
-      .oo-right-button {
-        @extend .oo-discover-btn;
-        right: 0;
-        text-align: left;
-      }
-    }
-
-    .oo-preview-panel {
-      width: 100%;
-      background-color: rgba(0,0,0,0.5);
-      position: absolute;
-      bottom: 0;
-      left: 0;
-      border-top: 1px solid $gray;
-      padding: 12px $panel-padding;
-
-      .oo-preview-caption {
-        font-size: 0.8em;
-        line-height:1.2em;
-        font-family: $font-family-roboto;
-      }
-      .oo-preview-text {
-        margin-top: 0;
-        font-size: 1.4em;
-        font-weight: bold;
-        font-family: $font-family-verdana;
-      }
-    }
-
-    .oo-disabled {
-      color: $gray;
-    }
-  }
-}
