+++ tests/components/moreOptionsPanel-test.js
@@ -42,11 +42,8 @@ describe('MoreOptionsPanel', function () {
       toggleShareScreen: function() {
         shareScreenToggled = true;
       },
-      toggleClosedCaptionScreen: function() {
-        ccScreenToggled = true;
-      },
       toggleScreen: function() {
-        qualityClicked = true;
+        toggleScreenClicked = true;
       }
     };
 
@@ -74,8 +71,7 @@ describe('MoreOptionsPanel', function () {
     for(var j=0; j<button.length; j++){
       TestUtils.Simulate.click(button[j]);
     }
-    expect(qualityClicked).toBe(true);
-    expect(ccScreenToggled).toBe(true);
+    expect(toggleScreenClicked).toBe(true);
     expect(shareScreenToggled).toBe(true);
     expect(discoveryScreenToggled).toBe(true);
   });
